+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 1.0
 - Use API proxy to avoid CORS problems
 - fix Dockerfile for production environments
+- Manage event types